refactor(editor): move whole-song save_song to the v2 API via a shared core
Consolidates the editor onto the v2 API (#1200): the primary whole-song save now runs
through api2.php (which enforces auth + the X-Requested-With same-origin CSRF guard),
instead of the legacy api.php which had session+role auth but NO CSRF.

Behaviour-preserving extraction (verified by diffing the original 1,094-line case body
against the new core — the ONLY changes are echo/http_response_code/break → return,
a $currentUser re-resolve in the catch for function scope, and an $action seed):
- NEW includes/manage/editor/save_song_core.php :: editorSaveSongCore(): array — the
  entire save (txn → UPSERT tblSongs → lyricLinesWriteComponents [rule #25] → credits →
  revision → SongCount → external-links → ISWC/Works auto-link → commit/rollback), now
  returning ['status','body'] instead of echoing. Self-sufficient (_songArtistsTableExists
  + placeColumnExists provided/required here so either API can call it).
- api.php case 'save_song' → thin shim (require core + echo) — full back-compat.
- api2.php gains case 'save_song' → ed2_respond(core) under the existing CSRF gate.
- editor.js autoSaveSongsPerSong now POSTs to api2 with X-Requested-With (so the save
  is CSRF-protected); request body (serialiseSongForSave) unchanged.
- Success response keeps the exact legacy wire shape (action = 'create'/'edit').
